Start of reorganizing
Chose same structure as Agda Stdlib but without the having the algebraic
structure modules parameterized over a set. This is because the `IsMagma` and
related records are helpful if the operator is already defined. We can't easily
make real numbers, meaning the functions that would normally be put into these
module parameters do not exist.